requirements.txt
discord.py
python-dotenv
gtts
beautifulsoup4
pytube
requests
watchdog